splitted objects.mjs and moved the gltf loading outside of the class
