+++ canvas_editor/editor/templates/editor/editor.html
@@ -12,6 +12,11 @@
       "navbar": "{% static 'js/editor/navbar.mjs' %}",
       "compass": "{% static 'js/editor/compass.mjs' %}",
       "objects": "{% static 'js/editor/objects.mjs' %}",
+      "canvasObject": "{% static 'js/editor/objects/canvasObject.mjs' %}",
+      "heliostat": "{% static 'js/editor/objects/heliostat.mjs' %}",
+      "receiver": "{% static 'js/editor/objects/receiver.mjs' %}",
+      "lightSource": "{% static 'js/editor/objects/lightSource.mjs' %}",
+      "terrain": "{% static 'js/editor/objects/terrain.mjs' %}",
       "command": "{% static 'js/commands/command.mjs' %}",
       "undoRedoHandler": "{% static 'js/commands/undoRedoHandler.mjs' %}",
       "singleObjectCommands": "{% static 'js/commands/singleObjectCommands.mjs' %}",
