+++ canvas_editor/static/js/editor/inspectorComponents.mjs
@@ -1,4 +1,5 @@
-import { CanvasObject } from "objects";
+import { CanvasObject } from "canvasObject";
+
 /**
  * Represents a single component of the inspector
  */
